Create Sidebar and update admin
