+++ KoiFengShuiSystemPRN212/KoiFengShuiSystem/AdminWPF.xaml
@@ -1,18 +1,21 @@
 ﻿<Window x:Class="KoiFengShuiSystem.AdminWPF"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:KoiFengShuiSystem"
-        mc:Ignorable="d"
-        Title="AdminWPF" Height="1000" Width="1850"
-        WindowStartupLocation = "CenterScreen">
+        Title="AdminWPF" Height="800" Width="1000"
+        WindowStartupLocation="CenterScreen">
     <Grid>
-        <Label Content="Admin Page" HorizontalAlignment="Center" Margin="0,45,0,0" VerticalAlignment="Top" Height="81" Width="238" FontSize="36"/>
-        <Button x:Name="btn_ManageUser" Content="Manage Users" HorizontalAlignment="Left" Margin="519,522,0,0" VerticalAlignment="Top" Height="95" Width="157" Click="btn_ManageUser_Click" FontSize="20"/>
-        <Button x:Name="btn_ManageKoi" Content="Manage Koi" HorizontalAlignment="Left" Margin="759,522,0,0" VerticalAlignment="Top" Height="95" Width="156" Click="btn_ManageKoi_Click" FontSize="20"/>
-        <Button x:Name="btn_ManagePond" Content="Manage Pond" HorizontalAlignment="Left" Margin="1021,522,0,0" VerticalAlignment="Top" Height="95" Width="157" Click="btn_ManagePond_Click" FontSize="20"/>
-        <Button x:Name="btn_logout" Content="Log Out" HorizontalAlignment="Left" Margin="1609,45,0,0" VerticalAlignment="Top" Height="85" Width="112" Click="btn_logout_Click" FontSize="20"/>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="200"/>
+            <ColumnDefinition Width="*"/>  
+        </Grid.ColumnDefinitions>
 
+        <local:Sidebar Grid.Column="0" Loaded="Sidebar_Loaded"/>
+
+        <!-- Main Content Area -->
+        <Grid Grid.Column="1">
+            <Label Content="Welcome to the Admin Page" HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="36"/>
+            
+        </Grid>
     </Grid>
 </Window>
