+++ KoiFengShuiSystemPRN212/KoiFengShuiSystem/Sidebar.xaml
@@ -1,0 +1,25 @@
+﻿<UserControl x:Class="KoiFengShuiSystem.Sidebar"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:KoiFengShuiSystem"
+             mc:Ignorable="d" 
+             Height="auto" Width="200">
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="200"/>
+            <ColumnDefinition Width="*"/>
+        </Grid.ColumnDefinitions>
+
+        <StackPanel Background="#2C3E50" VerticalAlignment="Stretch">
+            <Image Source="/hinh-anh-ho-ca-koi-dep_111108115.jpg" Width="116" Height="93" Margin="10"/>
+            <TextBlock Text="Admin Page" Foreground="White" FontSize="20" Margin="10" TextAlignment="Center"/>
+            <Separator Margin="10"/>
+            <Button x:Name="btn_ManageUser" Content="Manage Users"  Background="Transparent" Foreground="White" Click="btn_ManageUser_Click" Margin="10"/>
+            <Button x:Name="btn_ManageKoi" Content="Manage Koi"  Background="Transparent" Foreground="White" Click="btn_ManageKoi_Click" Margin="10"/>
+            <Button x:Name="btn_ManagePond" Content="Manage Pond"  Background="Transparent" Foreground="White"  Click="btn_ManagePond_Click" Margin="10"/>
+            <Button x:Name="btn_logout" Content="Log Out"  Background="Transparent" Foreground="White" Click="btn_logout_Click"  Margin="10"/>
+        </StackPanel>
+    </Grid>
+</UserControl>
