+++ KoiFengShuiSystemPRN212/KoiFengShuiSystem/UserManageWPF.xaml
@@ -2,14 +2,23 @@
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="User Management" Height="800" Width="1000"
+        xmlns:local="clr-namespace:KoiFengShuiSystem"
         WindowStartupLocation="CenterScreen" Loaded="Window_Loaded">
-    <Grid Background="#E8F0FE" Margin="10">
-        <Grid.RowDefinitions>
+
+    
+        <Grid Background="#E8F0FE" Margin="10">
+                 <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="200"/> <!-- Sidebar -->
+            <ColumnDefinition Width="*"/>   <!-- Nội dung -->
+        </Grid.ColumnDefinitions>
+        <local:Sidebar Grid.Column="0"/>
+        <Grid Grid.Column="1" Margin="10">
+            <Grid.RowDefinitions>
             <RowDefinition Height="Auto"/>
             <RowDefinition Height="*"/>
             <RowDefinition Height="Auto"/>
         </Grid.RowDefinitions>
-
+    
         <Label Content="MANAGER USER" HorizontalAlignment="Left" FontSize="24" FontWeight="Bold" Background="#FF6F61" Foreground="#FFFFFF" Padding="10" Margin="334,10,0,231" BorderBrush="#FF6F61" BorderThickness="1" Width="208"/>
 
         <StackPanel Orientation="Vertical" Margin="10,55,30,184" Grid.RowSpan="2">
@@ -72,5 +81,6 @@
             <Button x:Name="btnDelete" Content="Delete" Width="100" Click="btnDelete_Click" Margin="5" Background="#FF6F61" Foreground="#FFFFFF" BorderBrush="#FF6F61" BorderThickness="1"/>
             <Button x:Name="btnClose" Content="Close" Width="100" Click="btnClose_Click" Margin="5" Background="#FF6F61" Foreground="#FFFFFF" BorderBrush="#FF6F61" BorderThickness="1"/>
         </StackPanel>
+        </Grid>
     </Grid>
 </Window>
